requirements.txt
torch>=1.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.64.0